+++ packages/calcite-components/src/components/accordion-item/accordion-item.e2e.ts
@@ -1,6 +1,16 @@
 import { newE2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
 import { describe, expect, it } from "vitest";
-import { accessible, renders, slots, hidden, themed, focusable, reflects, defaults } from "../../tests/commonTests";
+import {
+  accessible,
+  renders,
+  slots,
+  hidden,
+  themed,
+  focusable,
+  reflects,
+  defaults,
+  t9n,
+} from "../../tests/commonTests";
 import { html } from "../../../support/formatting";
 import { CSS, IDS, SLOTS } from "./resources";
 
@@ -219,10 +229,16 @@ describe("calcite-accordion-item", () => {
     expect(headerContent.getAttribute("aria-expanded")).toBe("true");
   });
 
+  describe("translation support", () => {
+    t9n("calcite-accordion-item");
+  });
+
   it("should emit expanded/collapsed events when toggled", async () => {
+    const messages = await import("./assets/t9n/messages.json");
     const page = await newE2EPage();
     await page.setContent(html`<calcite-accordion-item heading="Test"></calcite-accordion-item>`);
     const item = await page.find("calcite-accordion-item");
+    const expandIcon = await page.find(`calcite-accordion-item >>> .${CSS.expandIcon}`);
 
     const expandSpy = await page.spyOnEvent("calciteAccordionItemExpand");
     const collapseSpy = await page.spyOnEvent("calciteAccordionItemCollapse");
@@ -232,11 +248,13 @@ describe("calcite-accordion-item", () => {
     expect(await item.getProperty("expanded")).toBe(true);
     expect(expandSpy).toHaveReceivedEventTimes(1);
     expect(collapseSpy).toHaveReceivedEventTimes(0);
+    expect(expandIcon.getAttribute("title")).toBe(messages.collapse);
 
     item.setProperty("expanded", false);
     await page.waitForChanges();
     expect(await item.getProperty("expanded")).toBe(false);
     expect(expandSpy).toHaveReceivedEventTimes(1);
     expect(collapseSpy).toHaveReceivedEventTimes(1);
+    expect(expandIcon.getAttribute("title")).toBe(messages.expand);
   });
 });