+++ packages/calcite-components/src/components/input-date-picker/input-date-picker.tsx
@@ -921,6 +921,11 @@ export class InputDatePicker
     if (this.shouldFocusRangeStart() || this.rangeStartValueChangedByUser) {
       return;
     }
+
+    if (this.proximitySelectionDisabled && this.valueAsDate[1] === null) {
+      return;
+    }
+
     this.open = false;
     this.focusInput();
   }