+++ packages/calcite-components/src/components/list-item/list-item.tsx
@@ -336,12 +336,6 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
    */
   calciteInternalListItemToggle = createEvent({ cancelable: false });
 
-  /**
-   *
-   * @private
-   */
-  calciteInternalListItemUpdateMoveToItems = createEvent({ cancelable: false });
-
   /** Fires when the close button is clicked. */
   calciteListItemClose = createEvent({ cancelable: false });
 
@@ -488,7 +482,6 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
   private handleSortHandleBeforeOpen(event: CustomEvent<void>): void {
     event.stopPropagation();
     this.calciteListItemSortHandleBeforeOpen.emit();
-    this.calciteInternalListItemUpdateMoveToItems.emit();
   }
 
   private handleSortHandleBeforeClose(event: CustomEvent<void>): void {